+++ src/backend/shortcuts/ipc_handler.ts
@@ -15,7 +15,6 @@ import {
 } from '../sideload/games'
 import { isMac } from 'backend/constants'
 import { notify } from 'backend/dialog/dialog'
-import { getWikiGameInfo } from 'backend/wiki_game_info/wiki_game_info'
 
 ipcMain.on('addShortcut', async (event, appName, runner, fromMenu) => {
   const isSideload = runner === 'sideload'
@@ -84,11 +83,9 @@ ipcMain.on('removeShortcut', async (event, appName, runner) => {
 
 ipcMain.handle('addToSteam', async (event, appName, runner) => {
   const gameInfo = getInfo(appName, runner)
-  const wikiInfo = await getWikiGameInfo(gameInfo.title, appName, runner)
 
   return addNonSteamGame({
-    gameInfo,
-    steamID: wikiInfo?.pcgamingwiki?.steamID
+    gameInfo
   })
 })
