+++ src/backend/shortcuts/nonesteamgame/constants.ts
@@ -5,11 +5,13 @@ const pictureExt = '.jpg'
 const coverArtSufix = 'p' + pictureExt
 const backGroundArtSufix = '_hero' + pictureExt
 const logoArtSufix = '_logo' + pictureExt
+const steamDBBaseURL = 'https://cdn.cloudflare.steamstatic.com/steam/apps'
 
 export {
   transparentSteamLogoHex,
   coverArtSufix,
   backGroundArtSufix,
   logoArtSufix,
-  pictureExt
+  pictureExt,
+  steamDBBaseURL
 }
